feat: enhance admin user console and login functionality
- Refactor user deletion and password reset logic in AdminUserConsole for better error handling and state management.
- Improve login form to support admin login via email prefix and streamline error handling for failed login attempts.
- Update AppShell navigation to include new access policies and relay nodes sections for admin users.
- Introduce new API endpoints for managing access policies and relay nodes in the admin API.
- Add utility functions for number formatting and key masking.
- Remove obsolete mock data file to clean up the codebase.
- Create a new Nginx configuration file for serving the application and proxying API requests.
- Update package dependencies and ensure compatibility with the latest versions.
- Extend TypeScript types to include new entities like AccessPolicy, RelayNode, and NodeAccount for better type safety.

+++ internal/db/log.go
@@ -14,10 +14,10 @@ type Log struct {
 	AccountID        uuid.UUID `gorm:"type:uuid;index" json:"account_id"`
 	Model            string    `gorm:"size:100;index" json:"model"`
 	IsStream         bool      `json:"is_stream"`
-	PromptTokens     int       `json:"prompt_tokens"`
-	CompletionTokens int       `json:"completion_tokens"`
-	TotalTokens      int       `json:"total_tokens"`
-	LatencyMs        int       `json:"latency_ms"`
+	PromptTokens     int64     `json:"prompt_tokens"`
+	CompletionTokens int64     `json:"completion_tokens"`
+	TotalTokens      int64     `json:"total_tokens"`
+	LatencyMs        int64     `json:"latency_ms"`
 	StatusCode       int       `json:"status_code"`
 	ErrorMessage     string    `gorm:"type:text" json:"error_message,omitempty"`
 }
