+++ internal/user/dto.go
@@ -87,10 +87,10 @@ type UsageLogItem struct {
 	CreatedAt        string `json:"created_at"`
 	Model            string `json:"model"`
 	IsStream         bool   `json:"is_stream"`
-	PromptTokens     int    `json:"prompt_tokens"`
-	CompletionTokens int    `json:"completion_tokens"`
-	TotalTokens      int    `json:"total_tokens"`
-	LatencyMs        int    `json:"latency_ms"`
+	PromptTokens     int64  `json:"prompt_tokens"`
+	CompletionTokens int64  `json:"completion_tokens"`
+	TotalTokens      int64  `json:"total_tokens"`
+	LatencyMs        int64  `json:"latency_ms"`
 	StatusCode       int    `json:"status_code"`
 	ErrorMessage     string `json:"error_message,omitempty"`
 }
